prints the size of various types on the computer it is compiled and run on

+++ 0x00-hello_world/6-size.c
@@ -17,11 +17,11 @@ int main(void)
 	long long int d;
 	float e;
 
-	printf("Size of a char: %d byte(s)\n", sizeof(a));
-	printf("Size of an int: %d byte(s)\n", sizeof(b));
-	printf("size of a long int: %d byte(s)\n", sizeof(c));
-	printf("size of a long long int: %d byte(s)\n", sizeof(d));
-	printf("size of a float: %d byte(s)\n", sizeof(e));
+	printf("Size of a char: %d byte(s)\n", (int) sizeof(a));
+	printf("Size of an int: %d byte(s)\n", (int) sizeof(b));
+	printf("size of a long int: %d byte(s)\n", (int) sizeof(c));
+	printf("size of a long long int: %d byte(s)\n", (int) sizeof(d));
+	printf("size of a float: %d byte(s)\n", (int) sizeof(e));
 
 	return (0);
 }
